requirements.txt
jinja2
python-dotenv
dbt-core
dbt-snowflake
